delete previous installs

+++ openshift4/baremetal/clean.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-rm -rf /root/ocp
+[ -d /root/ocp ] && rm -rf /root/ocp
 export LIBVIRT_DEFAULT_URI=$(grep libvirtURI install-config.yaml | sed 's/libvirtURI: //' | xargs)
 cluster=$(yq r install-config.yaml metadata.name)
 bootstrap=$(virsh list --name | grep "$cluster.*bootstrap")
